Refactoring of Code with coding convention. Removed old Event Bus

+++ Prototype/Source/Prototype/Messages/MsgHealthRefill.h
@@ -8,5 +8,5 @@ struct FMsgHealthRefill
 	GENERATED_USTRUCT_BODY()
 
 	AActor* TriggerActor;
-	int32 Amount;
+	float Amount;
 };
